Refactored proxy to use IProxy instead of CommonProxy
